Switch to an intermediate representation of a person (inc. memberships)
The motivation for this is partly that we want a simple representation
of everything we know about a person to put in the 'versions' array of a
person's PopIt record to enable easily reverting to that.

The representation we're now using is described in the comments at the
top of update.py.  That file also provides mixins for:

 - Taking that representation of a person, and updating a PopIt instance
   to match (PersonUpdateMixin) and
 - Fetching a person and their memberships from PopIt and turning it
   into that representation (PersonParseMixin)

This commit also rewrites the views that update data in PopIt to use
this representation.  (The constituency page view is still to be
updated, but should use PersonParseMixin in the future, in particular so
that it's easy to classify people into "known not to be standing again"
and "no known not to be standing again".)