Add a new 'parties' property to PopItPerson
This returns a dictionary showing the candidate's party memberships in
both 2010 and 2015, to replace the old 'party' property (which
essentially picked one randomly if they were different). This also
doesn't rely on the _update_organizations, which we want to remove.

Also remove the now unused get_year_party_map

Pair-programmed with Sym Roe (@symroe)