Match up Electoral Commission parties with YNMP parties where possible
This is done by matching names after some simple mangling into a key
(downcasing, removing non-alphanumeric characters, etc.)

For the few parties that couldn't be matched like this, we create them
but with a different ID scheme, which looks like ynmp-party:1234]
indicating that the party ID in the YourNextMP dump was 1234.  If the
party could be identified against an Electoral Commission registered
party, the ID will be something like party:5678, where 5678 is the
numeric part of the Electoral Commission party for that party.

This means that two pseudo-parties are created:

  'Independent' (ynmp-party:2)
  'Speaker seeking re-election' (ynmp-party:12522)

The other unmatched cases are these single constituency errors:

  British Unicorn Party
  Radical Reform Group
  New Independent Conservative Chelsea and Fulham
  Neath Port Talbot Independent Party
  Scottish Senior Citizens Unity Party
  National Liberal Party, The Third Way