Fix reverting to the original version of person's data
When you revert someone marked as standing in 2015 to the original version
of that person, the code will PUT a person's data where the value for the
'standing_in' attribute will just have a '2010' key rather than both a
'2010' and '2015' key. Unfortunately, due to:

  https://github.com/mysociety/popit-api/issues/95

... this will leave the person with an inconsistency between their
memberships and 'standing_in'. The same is true of 'party_memberships'.
As a workaround, first PUT a version of the person that will have null
values for 'standing_in' and 'party_memberships'.