Make sure the party_memberships field is actually written back to PopIt
When switching to doing the PopIt updates in update.py, I accidentally
left out the party_memberships field, and only included standing_in.
This doesn't actually make a difference to the running of the site (since
party membership data is in memberships anyway), but it's helpful to
people getting data from the API to have a simplified representation.