Allow fields to be cleared by setting them to empty
This is another effect of PopIt's non-standard behaviour in response
to PUT requests at the moment. (See
https://github.com/mysociety/popit-api/issues/95 )

We can fix this by making sure that simple fields are set to null if
they're not specified, and clearing links and contact details by setting
them to [] in the "purging" PUT before the real one.  sigh.

Fixes #33