Add templates for insisting on a source for candidacy create / delete
Because I was testing in a browser that used the HTML5 client-side
validation of the 'required' attribute, I missed that if you browser
ignores that, and you forget to add a source, you'll get an error
page because there was no template for that view.

This commit adds context data and templates for those views, so it
works even with older browsers.

Fixes #45