Add details of the license the data may be used under
Since this data includes the 2010 YourNextMP data (released under
CC-BY-SA) and we haven't heard from Edmund yet about whether an
alternative license would be possible, we should release it under a
compatible license: therefore I've said that the data can be used under
CC-BY-SA 4.0.

However, to leave the door open to change to a more liberal license
in the future I've also added the condition that data that people
submit to the site may be released under any of CC-BY-SA 4.0, CC0 or
ODbL depending on what we decide in the future and what Edmund thinks
would be compatible. In the meantime, the data will be explictly
available under CC-BY-SA 4.0.

This is a partial fix for #62, although as described above, it needs a
bit more discussion.