Add simple 'All time' and 'In the last week' leaderboards
This could do with some design attention (they're just two plain tables
at the moment) but adds leaderboards of users based on the top 10 number
of edits for "all time" and "in the last week".

This may not be the right incentive, so if this seems to be having a bad
effect, we might want to change this to hide the number of edits, as Tim
Green (@tfgg) suggested.

Fixes #66