When we don't have a constituency for 2010/2015, be more helpful
Unless we knew the candidate to be standing at the 2010 or 2015
election, the entry for that year on their candidate page would just be
blank. In fact, this could mean one of several things, e.g. "no
information at all", "won't stand", etc. This commit introduces
a templatetag that deals with this logic and simplifies the template.

Fixes #157