Insist that person IDs only consist of digits and update tests
The test data still used slugs as IDs, so this commit changes that, and
makes sure that the IDs in the test data match those in the current
YourNextMP database, just to avoid confusion.

A couple of tests needed their mocks updated as well; previously they
would pass a person ID which was an automatically generated mock return
value, but now that it's required to be only digits, the reverse creates
an error. This means setting mock_popit's return value for persons to
FakePersonCollection in a couple of extra places.