Add a (superuser only) option to merge two people
We're making this superuser only, since it's not trivial to revert a
merge in the same way that reverting a change to a field of a person
is.

When you go to a particular person's edit page, that person is
considered the 'primary' person, and the ID that you put into the
edit box is the 'secondary' person. In the case of any conflict
between the data, the field for the primary person is used.

The versions array of the primary person is prepended to the versions
array for the second person, and before them a "merged" version is
prepended. This isn't ideal - really, now that we support merging
you'd want a directed-acyclic graph of versions (like git, say) but I'm
reluctant to reinvent git's data model in a PopIt field.  As it is,
this might be confusing because the order of the versions is no longer
strictly reverse chronological; the first in the versions array will
still always be the most recent, however.