Add an NGINX_SSL option, to fix the results of is_secure
This setting will cause the following Django settings to be set:

  SECURE_PROXY_SSL_HEADER = ('HTTP_X_FORWARDED_PROTO', 'https')

... which will mean that request.is_secure will return correct results
when behind Nginx for TLS/SSL termination (so long as Nginx is
configured to set X-Forwarded-Proto). It will also tell django-allauth
to generate https URLs by default.

This should only be enabled when all traffic is proxied via Nginx
in this way.