Fix the insertion of a conflicting name into other_names
The logic for this was slightly wrong, but only because apparent with
dicts of particular size, where 'name' was dealt with before
'other_names'. The alternative name would be added to
result['other_name'], but when 'other_names' is dealt with, the value of
secondary['other_name'] rather than result['other_name'] was used as the
basis for the merge.