Fix the alignment of arrows under the standing / not standing buttons
The first-of-type selectors no longer worked, since another div was
introduced before the two .source-confirmation ones. This commit instead
overrides the left property for .source-confirmation classed divs after
the first one. It's still not great, because variability in the width of
the image on the left will cause the arrows not to be quite in the
middle of the button. @zarino pointed out better way of doing it:

 <zarino> that method of positioning the arrows is so hacky
 <zarino> I could do something much cleaner with hindsight
 <zarino> (make the arrow part of the *button*, not the yellow box, and
 just toggle a class on the button that makes the arrow below it appear
 or disappear – has the advantage that the arrow will always be centred
 below the button)
 <zarino> (wouldn't however work when the row of buttons breaks onto two
 lines, but that's pretty rare)

... but I'm commiting this incremental improvement until there's time to
do that proper fix.

Fixes #156