Don't overwrite values that look as if they've been manually corrected
The current policy is to overwrite any value in PopIt with the data from
the official PPC data sources. However, I'm worried that there might be
cases where the official data is wrong and has been manually corrected.

This commit changes the behaviour to check if the new value has been in
a previous version - if so, it doesn't overwrite a different current
value.