Introduce a demographics section & move gender information there
We want to make the gender information less prominent (since some people
had thought that odd, and it wasn't a conscious decision), so I've moved
it to a "Demographics" section further down the page. This also puts the
date of birth in "Personal Details" and adds the age in years
(calculated from the date of birth) into the "Demographics" section as
well.