Special case the ordering of some party emblems on import
At the moment the first image to be uploaded for a party is the one
that's made available though the image property, and the one that's used
by default. In some cases this means that you end up in a
region-specific version of the logo. This commit adds some special cases
for popular parties to make sure a useful general purpose logo is being
used.