Quick fix for a bug that broke updates for people with a DOB
The data returned from get_person() was originally just our reduced
representation, intended for recreating people from the versions array,
but recently I'd added some extra information that was only useful for
rendering in the template: date of birth as a datetime.date and
last_party.

This all needs to be rethought, but as a quick fix to enable update of
people with dates-of-birth again, I've changed get_person to return both
the reduced representation and also a dictionary of values intended just
for the template.