Make sure values are UTF-8 encoded before passing to writerow
Output of the data for all 2015 candidates was failing when output was
redirected to a file because some values were Unicode characters outside
the 7-bit ASCII range. This commit adds a test case with accented
characters and makes sure that all values in a row are encoded to UTF-8
before being passed to DictWriter.writerow.