Add standing again for the same / different party statistics
The statistics on the 'Numbers' page didn't obviously add up (see #185)
and the code that calculated them was a bit difficult to follow if you
wanted to work out why.  This commit hopefully makes their calculation
a bit clearer and also adds statistics to show the breakdown between
candidates who are standing again for the same party or a different party.