Avoid an occasional error on displaying person-create actions
When a new person is being created, we create the LoggedAction before
the interaction with PopIt happens (so that's logged even if there's an
error in the interaction with PopIt) and then set the popit_person_id
on it after we get that back from PopIt.  Unfortunately if someone
loaded the front page or /recent-changes between the creation and
update, that would be an error.

This commit fixes those error, by checking if there's a popit_person_id
before trying to get the person's URL.