Add a basic page listing all parties in PopIt / OEC
We could also link to these from the numbers-by-constituency page, but
initially I want to keep the party pages just for admininstrators, so
this is just an unadvertised index page.

The reasoning for keeping these pages somewhat hidden at first is that
they're not nicely designed, and I have some performance concerns for
parties with many members; it might be worth setting a fairly long cache
TTL for these pages.