Add middleware to handle errors better if PopIt is down
If PopIt is being restarted, then we currently get a flurry of error
emails and the users hitting the site then see an ugly Internal Server
Error page. This middleware catches ConnectionError (from requests)
and HttpServerError 503 (from slumber) and presents an error page in the
site's style saying the service is temporarily down.

PopIt and YourNextMP are being indepedently monitored by Nagios, so
hopefully this shouldn't affect our awareness of the site being down.

Fixes #126