Catch 404 exceptions from PopIt and make sure 404 is returned
We were getting error emails whenever a crawler tried to fetch a page
for a person who had been deleted, which was annoying. This was because
when slumber got a 404 from PopIt, it would raise a HttpClientError
exception. If we catch this and raise one of Django's Http404 exceptions
we shouldn't get an error email.

Fixes #211