Make the choices of why photos are allowed mutually exclusive
People seemed to be sometimes checking all of the boxes without
understanding what they meant, since some are mutually exclusive.

This commit switches the BooleanField fields in the model for why a
photo should be allowed to be used for a CharField with defined
choices.

It also removes use of the phrase "public domain" from the text
the user sees, since that's apparently widely misunderstood.