Photo review: allow the moderator to pick a different reason
It turns out that people submitting images often pick a reason for
it being allowed on the site which isn't consistent with their free
text explanation. As a result, it's helpful for the moderator to
be able to pick a different reason when approving it.

This commit introduces that change, and records both the user and
moderator's reasons in PopIt.