Add a command for fixing bad memberships data
There are some candidates in the database who have ended up with bad
membership data: the have no post membership corresponding to the
information in the 'standing_in' value.

This commit introduces a command that will recreate the PopIt
memberships for that person based on the data in standing_in and
party_memberships.