Add a "make this the primary image" option when reviewing a photo
Sometimes it's clear that the uploaded image is a better one to
use for the candidate, while sometimes it's just an extra one that
it might be nice to have.  Previously, all approved photos were
appended to the images array, which meant they wouldn't be used.
This commit changes that, so that now there's a "make this the
primary image" option, which, when selected, will insert that image
at the front of the images array.