Display messages from the Django messages framework
We'd like to use the Django messages framework to display "flash"
messages after actions to indicate that they worked, or to add
a further call-to-action.

This commit displays any messages from the messages framework in
the base template for the site. The immediate effect of this is
that messages relating to the auth system are suddenly displayed,
but I'll suppress them in later commits.

I actually took the CSS for this from an old pull request by
Sym Roe (@symroe), so thank-you for that.