For the moment, only display photo review messages
The reason I'm not just moving the message display wholly into the photo
review list template is that it would be good to use messages more
widely, e.g. for calls-to-action, but I don't have time to get the CSS
right for the contexts of all the other pages at the moment.  So for the
moment, just exclude all messages that don't have the photo-review tag.