Warn photo uploaders if there are already queued photos of that person
As suggested by @gerv, it's a good idea to warn someone uploading a
photo if there's already a photo of that person in the queue, so they
know whether it might be a waste of time.

This may also help the substantial number of people who don't realise
that there's a moderation queue - people often upload the same photo
several times in a row, presumably because they're confused that it
hasn't appeared yet.

Fixes #297