Switch to using Memcached for caching posts data
This is to resolve the problem that if you edited a person, their posts
would only be invalidated for one worker (at the moment there are
typically 5 processes handling YNMP requests).  Memcached is a single
cache shared between the workers, so switching to that as the cache for
posts should fix this.