Invalidate all cached posts associated with the person, not just one
There can be more than one different post in a person's memberships,
e.g. if they change constituency from one election to another.
Previously the code would only invalidate one of these posts because
the person -> post mapping in the cache was one-to-one.

This commit changes this so that there's no mapping between people and
posts maintained in the cache; instead the calling code has to work out
which posts to invalidate.