Remove any dissolved parties from /numbers/parties
We've made an effort recently to get rid of all the parties that are now
dissolved, and we now only consider the current parties when counting
numbers of candidates for each party. However, there will still be rows
in the cached_counts_cachedcount table that refer to now-dissolved
parties.

This commit introduces code to purge any such dissolved parties from the
cached_counts_cachedcount table when cached_counts_create is run.

Fixes #260