Add a command to create joint parties based on party descriptions
It's possible for candidates to stand for more than one party. We
believe this is only possible in cases where two different parties have
descriptions that refer to the other party and a shared joint
party name. For example:

 Solidarity - Scotland's Socialist Movement (party:604)
   has a description:
     Solidarity - Trade Unionist and Socialist Coalition (Joint Description with Trade Unionist and Socialist Coalition)

... and:

 Trade Unionist and Socialist Coalition (party:804)
   has a description:
     Solidarity - Trade Unionist and Socialist Coalition (Joint Description with Solidarity - Scotland's Socialist Movement)

So the command introduced by this commit would create a new party:

  Solidarity - Trade Unionist and Socialist Coalition

... with the ID party:604-party:804, and similar for any other pairs of
parties.

Fixes #168