Fix for approved photos not appearing
When someone approved a photo, we were correctly uploading the photo,
but then when saving the person to PopIt (to record a new version) it
was using the person's data from before the upload, which wouldn't have
the new photo in the 'images' array.  This commit makes sure that we
refetch the person's data from PopIt before recording the new version
and writing it back to PopIt.