Highlight candidates in constituency list for results recorders
This is to help people who are recording results to see easily that the
button they're pressing to mark a winner is associated with the
candidate they think it is. (It's not very subtle, but I think this
might help tired people not to make mistakes.) The CSS for people who
aren't in the "Results Recorders" group is unchanged.