Add a (slow) page to show a list of declared and undeclared constituencies
This is a quick hack at the end of a long sleepless election night, to
make sure we have some simple view of which constituencies there are
still to do.

Unfortunately, because it iterates over all posts with an embed of
'membership', this is slow to load, but it's better than nothing at this
stage.