Add an option to turn off edits for non-superusers
This is a pretty thoughtless implementation - it just deals with edits
from the update person page, and not merges, photo uploads and so on.
However as a (very tired) measure just to mean that we don't have to
worry about checking the recent changes now the election is over, this
is probably good enough.