Update Javascript for hiding and showing post and party drop-downs
When editing a person the post (constituency) drop-down and party
drop-downs should only be shown when the person is marked as standing.
Also, the party drop-down that's shown should be in the right "party
set" for the post.  This broke since we changed this page so that
candidacies for multiple posts could be shown and edited.

In addition, this adds a simple party set for Argentina, where all
parties are in the 'nacionale' party set.