Restructure the loading of (possibly country-specific) static data
This commit addresses two problems:

 * The code for mapping to party sets and post groups was
   election_specific but there was no natural way to implement this
   custom code, except for adding more and more functions to
   election_specific.py, which is messy.

 * Those election_specific.py functions were interdependent with the
   static_data.py data but defined on a different import path which
   meant it was easy to create knotty circular dependency errors.

This commit changes the MapItData and PartyData classes from essentially
just namespaces to classes that are intended to be instantiated in order
to get the data.  They should be subclassed by each election in order to
provide custom behaviour.

This also adds an AreaPostData for any data mappings that might rely on
both MapIt and Party data.

Hopefully this will be easier to understand and deal with.