Make the constituency detail view generic
The constituency detail view is the one you get from:
/constituency/{post_id}/{ignored_slug} in the UK or
/election/{election}/post/{post_id}/{ignored_slug} more
generically.  This view contained UK-specific functionality,
however, e.g. adding the links to electionleaflets.org, and
relying on the post ID for an MP being the same as the MapIt
ID of the constituency's area.

This commit moves the UK-specific functionality into the
elections/uk_general_election_2015 application, and removes
the UK-specific assumptions.