Allow both 'Primary Candidate' and 'Candidate' as membership roles
Tony Bowden suggested that a reasonable way of modelling the structure
of primary elections that decide the actual candidates for a post would
be:

  Primary election:

   - Candidates have a membership of the post with role 'Primary Candidate'
   - The winner of the primary election has a new membership with role
     'Candidate' created

  Actual election:

   - Candiates have a membership of the post with the role 'Candidate'
     (the membership created by recording the winner of the primary)
   - The winner of the actual election has a new membership with role
     'Member' created.

To support this, we change the ELECTIONS setting to include the expected
membership roles of candidates and winners, and the organization that
the post should be associated with.

We also have to update the code that filters memberships for candidate
memberships and records winners.