Support either a hidden or usable post field on the person form
Previously, the post field (previously "constituency field") was only
ever a hidden field, since its value was always known - the add person
field was embedded in the area or post page, or it was brought up when
complaining about validation errors, in which case the field's value
was carried over.  Now we want to be able to link to the 'person-create'
page to allow people to add a new candidate to any post in an election
so we add a 'hidden_post_widget' keyword argument to indicate that the
form is being embedded with a supplied post_id.

Fixes #419