Package upgrades: Django to 1.8.3, allauth to 0.21.0 and others
The motivations for this upgrade are two-fold:

* We need a more recent version of allauth to get translatable
  localizable versions of its text, as pointed out by Matthew
  here:

    https://github.com/mysociety/yournextrepresentative/issues/353#issuecomment-122213233

* We should be on Django 1.8 since it's a Long-Term Support
  release and 1.7 will stop getting security updates at some point
  soon.

These upgrades also requires some other upgrades either as direct
dependencies or in order to get rid of new warnings and errors,
in the case of:

  * django-debug-toolbar
  * django-nose

We also needed to remove South, which is no longer used, and was
generating warnings.

Automatically migrating to this version will fail, since allauth now has
Django migrations rather than South migrations and it doesn't know that
various relations have already been created.  You will need to fake two
of these migrations with:

  ./manage.py migrate --fake account 0001
  ./manage.py migrate --fake socialaccount 0001_initial

In order to suppress a deprecation warning about supplying both
'auto_now_add' / 'auto_now' and 'default' for DateTimeField fields
on QueuedImage, this also includes a migration to remove the 'default'.