Make the tests network-independent again
This finishes the work of mocking and patching network-dependent
functionality in tests so that they don't rely on any external APIs.

As an explanatory note, I've removed one of the examples that uses
SW1 1AA (Buckingham Palace), since the constituency that contains that
postcode (Cities of London and Westminster, 65759) wasn't in the
fake data returned for get_all_posts with these fixtures.  It seemed
easier to change the postcode than to add yet another constituency to
that data.

Fixes #448