Remove an unused method with a confusing comment
The tidy_address method isn't used anywhere. The comment suggests that
it should be used for adding, say, ", Argentina", to an address before
geocoding, but in fact this is done by setting the 'country' class
variable on the AddressFinderView subclass.  So, remove this to avoid
confusion.

Fixes #482, which is evidence of this confusion ;)