BF: add extra fields for Burkina Faso candidates
For Burkina Faso, it's been requested that we add a "CV" and "Program"
fields for candidates who are standing in the presidential election.

Since any candidate might be edited to make them a candidate in the
presidential election, add these fields to all candidates in the
Burkina Faso site for the moment.

If it's really required that these fields are dynamically hidden and
unhidden as someone is marked as standing or not standing in the
presidential election, we can deal with that later.