Add support for party lists (for PR electoral systems)
If you set the 'party_lists_in_use' option in election data to True,
then you can edit the position of candidates in their party list and
the candidates will be shown grouped by parties.

This commit also adds a party list page for showing just the candidates
from a particular party for a post - if they have numbered positions in
that list they'll be shown in that order on that page.

The design and CSS changes in this commit are by Zarino Zappia (@zarino).