Fix so that the fingerprinted djangojs.js files can be found
The "./manage.py compilejsi18n" command just writes the djangojs.js
files under STATIC_ROOT rather than being processed by
staticfiles, so the fingerprinted version of this file is never
created. It seems that to get staticfiles to process the output of
compilejsi18n you have to follow one of these approaches:

  https://django-statici18n.readthedocs.org/en/latest/faq.html#how-to-configure-static-files-with-django-statici18n

Since we already have mysite/static added to STATICFILES_DIRS this
commit takes the latter approach and gets compilejsi18n to output
to mysite/static/jsi18n/...

We also alter the order of the commands in bin/pre-deploy so that
these files are generated before collectstatic is run.

Fixes #525