Fix errors caused by dissolved parties not being in PARTY_DATA
Once a party is dissolved, we no longer list it in
all-parties-from-popit.json, so looking it up in
PARTY_DATA.party_id_to_name fails.

This is only to get the party name, and we can get that via alternative
means - e.g. the party name via party_memberships or the person's
memberships.

Fixes #541