Fix dreadful performance loading long area pages when DjDT is active
Django Debug Toolbar seems to make page loads incredibly slow if the
context contains data that's very large when rendered.  In this case
that's the post data with multiple levels of embedding, so this
commit just makes sure the 'memberships' element is stripped once
it's no longer useful and before it's passed to the context.