Posts need an 'elections' array attribute, not an 'election' string
The previous change (e9a502a0b) to store an 'election' attribute on
each post actually wasn't quite right, because people can be elected
to the same post in more than one election.  (The example we have at
the moment is that the UK site has both a 2010 and 2015 election
with exactly the same posts.)  The previous change meant that the
UK general election posts only ended up associated with one election
when they should be returned from searches for both.

This commit changes that, so that posts are expected to have a
'elections' array of election IDs instead of an 'election' string
with a single election ID.

After deploying this, you will need to run:

  ./manage.py candidates_create_popit_posts

... to make sure all the posts have the correct 'elections' attribute.