Add migrations and model updates to move from PopIt to django-popolo
[This commit represents joint work by Struan Donald (@struan) and Mark
Longair (@mhl20), squashed together from many different commits.]

PopIt is being deprecated, so we have decided to switch this project
to use the django-popolo Django application and its models, which can
also store data using the Popolo data model.

The most significant data migration is
candidates.0009_migrate_to_django_popolo which, if ELECTION_APP is set
to one of the known current installations, will fetch data from the
live PopIt instance for that installation and import that data to the
djang-popolo models.

django-popolo's models will not store all of the information
associated with People, Organizations, Memberships and Posts, however.
We've augmented them via a series of models (PersonExtra, PostExtra,
etc.), each with a OneToOneField to the parent django-popolo class in
order to add additional data.  (This is done with explicit
OneToOneField relationships rather than the implicit OneToOneFields
from multi-table inheritance so we have more control over which JOINs
are done, in case of performance issues associated with multi-table
inheritance.)  These extra models add the following:

 * Multiple images per Person and Organization, with metadata to
   indicate who updated them and what their copyright status is.

 * Previous versions of candidates, serialized to JSON.

 * The electoral register that an a party (Organization) is on. (This
   could also have been done with the scheme of the identifier of the
   party on that register, but it's convenient to have as a distinct
   field.

 * slug fields on the Post and Organization models, which are used for
   non-integer primary keys.

 * 'cv' and 'program' fields for people, required by the Burkina Faso
   site. (In the future we will want to support additional
   country-specific qfields more generically, but this migration is an
   as-direct-as-possible mapping onto django-popolo.)

 * The election that a membership representing a candidacy relates to.

 * The elections that a post exists in.

 * A property of posts that indicates that they're locked (meaning
   that edits of people standing for that post are restricted).

 * A 'group' property of posts, used in the UK to say which region of
   the country the post is in (e.g. Scotland, Wales, etc.)

 * Party sets to which parties can belong - this is used to provide a
   restricted list of parties that can be chosen for people based on
   the constituency they're standing in.

 * An 'elected' property for candidacy memberships to indicate whether
   the person was or wasn't elected to that post at the
   election. (It's None if that's not known or this membership doesn't
   represent a candidacy.)

 * A 'party_list_position' property on candidacy memberships for when
   a party list system is in use to indicate the position of that
   candidate on their party's list.

 * A representation of the type of a particular area.

 * A 'not_standing' many-to-many relationship between people and
   elections to indicate which elections someone is known not to be
   standing in.

There are various existing models (e.g. QueuedImage, LoggedAction,
etc.) which refered to PopIt IDs or election slugs, which are now
migrated to have foreign keys to the new Django models.

This commit also removes the MaxPopItIds model, which was tracking
the maximum person ID in PopIt - this is no longer needed.