Update translation strings and formatting of email bodies
There are a few places where YNR emits email (e.g. reporting the
results of image moderation & alerts on rejected edits) - the
templates for these emails meant that fixed text formatting would
appear as msgid strings to be translated, which is confusing for
the translators and in viewing diffs.

In this case we can't just use {% blocktrans trimmed %} in the
email templates, because we want to hard-wrap the lines of these
translated emails at about 64 characters using the wordwrap
template filter; in Django 1.8 this means passing in the
translated strings in the context for template rendering.