Standardise on all msgid strings being single lines
Translated strings that we fetch from Transifex with the Transifex
command line client are formatted as a single line; I think it's
easiest to review changes if the msgid lines are also a single
line (i.e. not wrapped).

This commit updates all msgid strings that were split across
multiple lines and also updates the documentation describing how
to run makemessages to add the --no-wrap option, so that the
generated messages are also single-line.

This will also make it easier for contributors who can't both push
and pull from Transifex to add msgids for translated strings
they've added, since the diffs after running makemessages with
--no-wrap should be minimal.