merge moderation tests into one file
This gets round a bug where setting the MEDIA_ROOT with
override_settings upsets the i18njs module and causes it not to be able
to find its settings. If we move the override to be at a class level
then it doesn't cause the bug, presumably because the issue is in the
restoring and then re-overriding the settings. Not a very satisfactory
fix but it wasn't clear how else to resolve it