Make '' the default value for "complex fields" via getattr, not None
This is for consistency with Django where CharField fields are
usually recommended to be empty string rather than None as the
null value - the 'complex fields' are all essentially strings,
usually URLs.

This also makes it easy to maintain the same serialization to JSON
used in the version history as before.