elected status can be None if no results
can't just check if a membership with elected=True exists as there is a
difference between it existing and being False ( not elected ) and
either not existing or exising and being None ( no results yet ).
Change get_elected to either return the value of elected or None if it
doesn't exist.