Remove any duplicate Membership objects after import
The PopIt instances backing YourNextRepresentative sites turn out to
have lots of duplicate memberships which represent exactly the same
relationships.  Someone *might* want these for a good reason, so I
don't think this code stripping out duplicates is worth sending
upstream, but it's needed for YNR so that we get a unique candidacy
person per election.