management command to create areas and posts WIP
Given the MapIt URL of a region, an area type and a post id format this
will add an area and a post, plus associated other entities, for each
area of the given type covered by the region.

It will also link the posts to all the elections present.

Currently assumes that all the elections are the same but for different
years and that all posts are part of all elections. This clearly is not
always going to be the case so needs some work.